extract Address into a new @helium/address package
