+++ integration_tests/tests/create_and_submit_payment.spec.ts
@@ -1,11 +1,11 @@
 import nock from 'nock'
-import { Keypair, Address } from '@helium/crypto'
+import { Keypair } from '@helium/crypto'
+import Address from '@helium/address'
 import { Client } from '@helium/http'
 import { PaymentV1, PaymentV2 } from '@helium/transactions'
 import { bobWords, bobBip39Words, aliceB58 } from '../fixtures/users'
 
 test('create and submit a payment txn', async () => {
-
   const bob = await Keypair.fromWords(bobWords)
   const aliceAddress = Address.fromB58(aliceB58)
 
@@ -57,7 +57,7 @@ test('create and submit a PaymentV2 txn', async () => {
   const signedPaymentTxn = await paymentTxn.sign({ payer: bob })
   const serializedTxn = signedPaymentTxn.toString()
   expect(serializedTxn).toBe(
-    'wgGOAQohATUaccIv7+wiMZNq0oJrIX7OOdn3f8bEljmSYpnDhpKVEiUKIQGcZZ1yPMHoEKcuePfer0c2qH8Q74/PyAEAtTMn5+5JpBAKIAEqQK88GjmG9CrESHVdcL//ZfWD+KsBnbKmZqKlx8oD89FUms7OjZNcL5NiQ4o0jREg+ahkjc2jX4SgKBBniM+QoAA='
+    'wgGOAQohATUaccIv7+wiMZNq0oJrIX7OOdn3f8bEljmSYpnDhpKVEiUKIQGcZZ1yPMHoEKcuePfer0c2qH8Q74/PyAEAtTMn5+5JpBAKIAEqQK88GjmG9CrESHVdcL//ZfWD+KsBnbKmZqKlx8oD89FUms7OjZNcL5NiQ4o0jREg+ahkjc2jX4SgKBBniM+QoAA=',
   )
 
   nock('https://api.helium.io')
@@ -75,7 +75,6 @@ test('create and submit a PaymentV2 txn', async () => {
 })
 
 test('using the bip39 checksum word should match serialization', async () => {
-
   const bob = await Keypair.fromWords(bobBip39Words)
   const aliceAddress = Address.fromB58(aliceB58)
 
@@ -94,6 +93,6 @@ test('using the bip39 checksum word should match serialization', async () => {
   const serializedTxn = signedPaymentTxn.toString()
 
   expect(serializedTxn).toBe(
-    'wgGOAQohATUaccIv7+wiMZNq0oJrIX7OOdn3f8bEljmSYpnDhpKVEiUKIQGcZZ1yPMHoEKcuePfer0c2qH8Q74/PyAEAtTMn5+5JpBAKIAEqQK88GjmG9CrESHVdcL//ZfWD+KsBnbKmZqKlx8oD89FUms7OjZNcL5NiQ4o0jREg+ahkjc2jX4SgKBBniM+QoAA='
+    'wgGOAQohATUaccIv7+wiMZNq0oJrIX7OOdn3f8bEljmSYpnDhpKVEiUKIQGcZZ1yPMHoEKcuePfer0c2qH8Q74/PyAEAtTMn5+5JpBAKIAEqQK88GjmG9CrESHVdcL//ZfWD+KsBnbKmZqKlx8oD89FUms7OjZNcL5NiQ4o0jREg+ahkjc2jX4SgKBBniM+QoAA=',
   )
 })
