+++ packages/address/src/index.ts
@@ -1,0 +1,10 @@
+/**
+ * [[include:address/README.md]]
+ * @packageDocumentation
+ * @module address
+ */
+
+export { default } from './Address'
+export * as NetTypes from './NetTypes'
+export * as KeyTypes from './KeyTypes'
+export * as utils from './utils'
