+++ packages/crypto-react-native/src/index.ts
@@ -6,9 +6,6 @@
 
 export { default as Mnemonic } from './Mnemonic'
 export { default as Keypair } from './Keypair'
-export { default as Address } from './Address'
-export * as NetType from './NetType'
-export * as KeyType from './KeyType'
 export * as utils from './utils'
 
 global.Buffer = require('safe-buffer').Buffer
