+++ packages/crypto/package.json
@@ -26,12 +26,11 @@
     "build": "yarn run clean && tsc"
   },
   "dependencies": {
-    "bs58": "^4.0.1",
+    "@helium/address": "^3.52.0",
     "create-hash": "^1.2.0",
     "libsodium-wrappers": "^0.7.6"
   },
   "devDependencies": {
-    "@types/bs58": "4.0.1",
     "@types/create-hash": "1.2.2",
     "@types/libsodium-wrappers": "0.7.8"
   },
