--- packages/crypto/src/KeyType.ts
@@ -1,9 +1,0 @@
-export const ECC_COMPACT_KEY_TYPE = 0
-export const ED25519_KEY_TYPE = 1
-
-export const SUPPORTED_KEY_TYPES = [
-  ECC_COMPACT_KEY_TYPE,
-  ED25519_KEY_TYPE,
-]
-
-export type KeyType = number
