--- packages/crypto/src/NetType.ts
@@ -1,9 +1,0 @@
-export const MAINNET = 0x00
-export const TESTNET = 0x10
-
-export const SUPPORTED_NET_TYPES = [
-  MAINNET,
-  TESTNET,
-]
-
-export type NetType = number
