+++ packages/crypto/src/index.ts
@@ -6,7 +6,4 @@
 
 export { default as Mnemonic } from './Mnemonic'
 export { default as Keypair } from './Keypair'
-export { default as Address } from './Address'
-export * as NetType from './NetType'
-export * as KeyType from './KeyType'
 export * as utils from './utils'
