+++ packages/http/src/resources/__tests__/Challenges.spec.ts
@@ -1,8 +1,8 @@
 import nock from 'nock'
-import { Address, NetType } from '@helium/crypto'
+import Address, { NetTypes } from '@helium/address'
 import Client from '../../Client'
 
-Address.fromB58 = jest.fn(() => new Address(0, NetType.MAINNET, 0, new Uint8Array()))
+Address.fromB58 = jest.fn(() => new Address(0, NetTypes.MAINNET, 0, new Uint8Array()))
 
 export const challengeFixture = (params = {}) => ({
   type: 'poc_receipts_v1',
