+++ packages/http/src/resources/__tests__/Transactions.spec.ts
@@ -1,5 +1,5 @@
 import nock from 'nock'
-import { Address, NetType } from '@helium/crypto'
+import Address, { NetTypes } from '@helium/address'
 import Client from '../../Client'
 import {
   UnknownTransaction,
@@ -11,7 +11,7 @@ import {
 } from '../../index'
 import { RewardsV2 } from '../../models/Transaction'
 
-Address.fromB58 = jest.fn(() => new Address(0, NetType.MAINNET, 0, new Uint8Array()))
+Address.fromB58 = jest.fn(() => new Address(0, NetTypes.MAINNET, 0, new Uint8Array()))
 
 describe('submit', () => {
   it('posts to the pending transactions endpoint', async () => {
