+++ packages/http/src/resources/__tests__/Validators.spec.ts
@@ -1,5 +1,5 @@
 import nock from 'nock'
-import { Address, NetType } from '@helium/crypto'
+import Address, { NetTypes } from '@helium/address'
 import Client from '../../Client'
 
 // eslint-disable-next-line import/prefer-default-export
@@ -85,7 +85,7 @@ export const rewardSumListFixture = () => ({
   ],
 })
 
-Address.fromB58 = jest.fn(() => new Address(0, NetType.MAINNET, 0, new Uint8Array()))
+Address.fromB58 = jest.fn(() => new Address(0, NetTypes.MAINNET, 0, new Uint8Array()))
 
 describe('get', () => {
   nock('https://api.helium.io').get('/v1/validators/fake-validator-address').reply(200, {
