+++ packages/transactions/package.json
@@ -27,7 +27,7 @@
     "build": "yarn run clean && tsc"
   },
   "dependencies": {
-    "@helium/crypto": "^3.60.0",
+    "@helium/address": "^3.52.0",
     "@helium/proto": "^1.5.0",
     "@types/libsodium-wrappers": "^0.7.8",
     "long": "^4.0.0",
