+++ packages/transactions/src/utils.ts
@@ -1,4 +1,4 @@
-import { Address } from '@helium/crypto'
+import Address from '@helium/address'
 import Long from 'long'
 import { Addressable } from './types'
