+++ typedoc.json
@@ -6,7 +6,8 @@
     "packages/http/src/index.ts",
     "packages/onboarding/src/index.ts",
     "packages/proto-ble/build/index.d.ts",
-    "packages/transactions/src/index.ts"
+    "packages/transactions/src/index.ts",
+    "packages/address/src/index.ts"
   ],
   "exclude": ["**/*.spec.ts"],
   "includes": "packages",
